+++ lib/group/replica/data.ex
@@ -327,14 +327,14 @@ defmodule Group.Replica.Data do
     end
   end
 
-  def pg_members(name, shard, cluster, key) do
+  def pg_members(name, shard, cluster, key, limit \\ :infinity) do
     table = pg_by_key_table(name, shard)
     # Use match spec to find all entries with the given {cluster, key, _pid} prefix
     match_spec = [
       {{{cluster, key, :"$1"}, :"$2", :_, :_}, [], [{{:"$1", :"$2"}}]}
     ]
 
-    :ets.select(table, match_spec)
+    select(table, match_spec, limit)
   end
 
   def pg_members_with_node(name, shard, cluster, key) do
@@ -347,14 +347,18 @@ defmodule Group.Replica.Data do
     :ets.select(table, match_spec)
   end
 
-  def pg_members_by_prefix(name, shard, cluster, prefix) do
+  def pg_members_by_prefix(name, shard, cluster, prefix, limit \\ :infinity) do
     table = pg_by_key_table(name, shard)
     prefix_end = next_binary_prefix(prefix)
 
-    :ets.select(table, [
-      {{{cluster, :"$1", :"$2"}, :"$3", :_, :_},
-       [{:andalso, {:>=, :"$1", prefix}, {:<, :"$1", prefix_end}}], [{{:"$2", :"$3"}}]}
-    ])
+    select(
+      table,
+      [
+        {{{cluster, :"$1", :"$2"}, :"$3", :_, :_},
+         [{:andalso, {:>=, :"$1", prefix}, {:<, :"$1", prefix_end}}], [{{:"$2", :"$3"}}]}
+      ],
+      limit
+    )
   end
 
   def pg_members_local(name, shard, cluster, key) do
@@ -366,6 +370,35 @@ defmodule Group.Replica.Data do
     ]
 
     :ets.select(table, match_spec)
+  end
+
+  def pg_members_local_with_meta(name, shard, cluster, key, limit \\ :infinity) do
+    local_node = node()
+    table = pg_by_key_table(name, shard)
+
+    match_spec = [
+      {{{cluster, key, :"$1"}, :"$2", :_, :"$3"}, [{:==, :"$3", local_node}], [{{:"$1", :"$2"}}]}
+    ]
+
+    select(table, match_spec, limit)
+  end
+
+  def pg_members_local_by_prefix(name, shard, cluster, prefix, limit \\ :infinity) do
+    local_node = node()
+    table = pg_by_key_table(name, shard)
+    prefix_end = next_binary_prefix(prefix)
+
+    select(
+      table,
+      [
+        {{{cluster, :"$1", :"$2"}, :"$3", :_, :"$4"},
+         [
+           {:==, :"$4", local_node},
+           {:andalso, {:>=, :"$1", prefix}, {:<, :"$1", prefix_end}}
+         ], [{{:"$2", :"$3"}}]}
+      ],
+      limit
+    )
   end
 
   @doc """
@@ -878,5 +911,15 @@ defmodule Group.Replica.Data do
     :ets.new(cluster_leases_table(name), set_opts)
 
     {:ok, %{name: name, num_shards: num_shards}}
+  end
+
+  defp select(table, match_spec, :infinity), do: :ets.select(table, match_spec)
+  defp select(_table, _match_spec, 0), do: []
+
+  defp select(table, match_spec, limit) when is_integer(limit) and limit > 0 do
+    case :ets.select(table, match_spec, limit) do
+      :"$end_of_table" -> []
+      {matches, _continuation} -> matches
+    end
   end
 end